Fix server components warning in examples page
